ui_state_capture_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ui_state_capture_agent"
version = "0.1.0"
description = "UI state capture agent"
requires-python = ">=3.10"
authors = [
  { name = "Softlight Team" }
]

dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "sqlalchemy==2.0.31",
  "psycopg2-binary==2.9.9",
  "minio==7.2.7",
  "openai==1.35.10",
  "playwright==1.45.1",
  "python-dotenv==1.0.1",
  "Jinja2==3.1.4",
  "httpx==0.27.0",
  "pydantic==2.8.2",
  "pydantic-settings==2.3.4"
]

[project.optional-dependencies]
test = [
  "pytest==8.2.2"
]

[tool.setuptools.packages.find]
where = ["src"]
